+++ sandbox/main.java
@@ -1,0 +1,9 @@
+// :snippet-start: c1
+hi
+// :snippet-end:
+// :snippet-start: c2
+bye
+// :snippet-end:
+// :snippet-start: c3
+should not snip
+// :snippet-end:
