--- src/bluehawk/actions/actions.ts
@@ -1,3 +1,0 @@
-export interface ActionArgs {
-  waitForListeners?: boolean;
-}
