+++ src/bluehawk/actions/index.ts
@@ -3,4 +3,4 @@ export * from "./copy";
 export * from "./listStates";
 export * from "./listCommands";
 export * from "./snip";
-export * from "./actions";
+export * from "./ActionArgs";
